fix: LongKeyMap 哈希函数改为 Long.hashCode() 混合高低 32 位
原始实现 Integer.hashCode((int)key) 会丢弃高 32 位，导致高
位不同的 long key 产生相同哈希，增加碰撞概率。改为标准
Long.hashCode()（高低 32 位异或），均匀分布哈希值。

+++ evrete-core/src/main/java/org/evrete/collections/LongKeyMap.java
@@ -53,7 +53,7 @@ public class LongKeyMap<T> implements Iterable<T> {
 
 
     private int hash(long key) {
-        return Integer.hashCode((int)key) & (table.length - 1);
+        return Long.hashCode(key) & (table.length - 1);
     }
 
     public T put(long key, T value) {
